Split out daily interval schedule  tests.
